port @truffle/compile-ligo to ts
