+++ packages/compile-ligo/.gitignore
@@ -1,0 +1,3 @@
+package-lock.json
+dist
+node_modules
