+++ packages/compile-ligo/README.md
@@ -1,2 +1,3 @@
 # @truffle/compile-ligo
+
 Ligo compiler interface for truffle.
