+++ packages/compile-ligo/package.json
@@ -5,16 +5,23 @@
   "author": "CruzMolina <7537712+CruzMolina@users.noreply.github.com>",
   "homepage": "https://github.com/trufflesuite/truffle#readme",
   "license": "MIT",
-  "main": "index.js",
+  "main": "./dist/src/index.js",
+  "types": "./dist/src/index.d.ts",
   "directories": {
-    "scripts": "scripts",
-    "test": "tests"
+    "doc": "./README.md",
+    "lib": "./dist",
+    "test": "./tests"
   },
+  "files": [
+    "./dist/src"
+  ],
   "repository": {
     "type": "git",
     "url": "git+https://github.com/trufflesuite/truffle.git"
   },
   "scripts": {
+    "build": "tsc",
+    "prepare": "yarn build",
     "test": "./scripts/test.sh"
   },
   "publishConfig": {
@@ -24,14 +31,24 @@
     "url": "https://github.com/trufflesuite/truffle/issues"
   },
   "devDependencies": {
-    "@truffle/config": "^1.5.0-tezos.0",
-    "mocha": "5.2.0"
+    "@types/minimatch": "^3.0.3",
+    "@types/mocha": "^8.0.0",
+    "@types/node": "^14.0.24",
+    "mocha": "8.0.1",
+    "ts-node": "8.10.2",
+    "typescript": "^3.9.7"
   },
   "dependencies": {
     "@truffle/compile-solidity": "^4.3.0-tezos.12",
+    "@truffle/config": "^1.5.0-tezos.0",
     "@truffle/contract-sources": "^0.4.0-tezos.0",
-    "async": "^3.1.0",
     "colors": "^1.4.0",
     "minimatch": "^3.0.4"
-  }
+  },
+  "keywords": [
+    "truffle",
+    "tezos",
+    "blockchain",
+    "ligo"
+  ]
 }
